Fixed unnnessary cast

+++ core/src/main/java/hivemall/math/matrix/builders/CSCMatrixBuilder.java
@@ -60,7 +60,7 @@ public final class CSCMatrixBuilder extends MatrixBuilder {
 
         rows.add(row);
         cols.add(col);
-        values.add((float) value);
+        values.add(value);
         this.maxNumColumns = Math.max(col + 1, maxNumColumns);
         return this;
     }
